backend/requirements.txt
Django==3.2
pytz==2021.1
sqlparse==0.4.1
django-colorfield==0.7.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-filter==2.4.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyJWT==2.1.0
requests==2.26.0
djoser==2.1.0
drf-extra-fields==3.4.0
pyflakes==2.5.0
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.8.6
